Fix 'Additional Users' Popover Display Logic for Profile Clicks (#270)
* Modify profile click functionality to only show 'Additional Users' popover when more than 4 users are present

* Separate UserPresence component from DocumentHeader component

* Fix code format
