+++ frontend/src/components/headers/DocumentHeader.tsx
@@ -4,8 +4,6 @@ import VerticalSplitIcon from "@mui/icons-material/VerticalSplit";
 import VisibilityIcon from "@mui/icons-material/Visibility";
 import {
 	AppBar,
-	Avatar,
-	AvatarGroup,
 	IconButton,
 	Paper,
 	Stack,
@@ -13,14 +11,8 @@ import {
 	ToggleButtonGroup,
 	Toolbar,
 	Tooltip,
-	Popover,
-	Typography,
-	List,
-	ListItem,
-	ListItemAvatar,
-	ListItemText,
 } from "@mui/material";
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useNavigate } from "react-router-dom";
 import { useList } from "react-use";
@@ -31,6 +23,12 @@ import { YorkieCodeMirrorPresenceType } from "../../utils/yorkie/yorkieSync";
 import DownloadMenu from "../common/DownloadMenu";
 import ShareButton from "../common/ShareButton";
 import ThemeButton from "../common/ThemeButton";
+import UserPresence from "./UserPresence";
+
+export type Presence = {
+	clientID: ActorID;
+	presence: YorkieCodeMirrorPresenceType;
+};
 
 function DocumentHeader() {
 	const dispatch = useDispatch();
@@ -46,12 +44,7 @@ function DocumentHeader() {
 			clear: clearPresenceList,
 			filter: filterPresenceList,
 		},
-	] = useList<{
-		clientID: ActorID;
-		presence: YorkieCodeMirrorPresenceType;
-	}>([]);
-
-	const [anchorEl, setAnchorEl] = useState<HTMLElement | null>(null);
+	] = useList<Presence>([]);
 
 	useEffect(() => {
 		if (editorState.shareRole === "READ") {
@@ -96,20 +89,6 @@ function DocumentHeader() {
 		navigate(`/${workspaceState.data?.slug}`);
 	};
 
-	// Display additional users in a popover when there are more than 4 users
-	const handleOpenPopover = (event: React.MouseEvent<HTMLElement>) => {
-		setAnchorEl(event.currentTarget);
-	};
-
-	// Display None additional users in a popover when there are more than 4 users
-	const handleClosePopover = () => {
-		setAnchorEl(null);
-	};
-
-	const popoverOpen = Boolean(anchorEl);
-
-	const hiddenAvatars = presenceList.slice(3);
-
 	return (
 		<AppBar position="static" sx={{ zIndex: 100 }}>
 			<Toolbar>
@@ -151,55 +130,7 @@ function DocumentHeader() {
 						<DownloadMenu />
 					</Stack>
 					<Stack direction="row" alignItems="center" gap={1}>
-						<AvatarGroup max={4} onClick={handleOpenPopover}>
-							{presenceList?.map((presence) => (
-								<Tooltip key={presence.clientID} title={presence.presence.name}>
-									<Avatar
-										alt={presence.presence.name}
-										sx={{ bgcolor: presence.presence.color }}
-									>
-										{presence.presence.name[0]}
-									</Avatar>
-								</Tooltip>
-							))}
-						</AvatarGroup>
-						<Popover
-							open={popoverOpen}
-							anchorEl={anchorEl}
-							onClose={handleClosePopover}
-							anchorOrigin={{
-								vertical: "bottom",
-								horizontal: "left",
-							}}
-						>
-							<Paper sx={{ padding: 2 }}>
-								<Typography variant="subtitle2">Additional Users</Typography>
-								<List>
-									{hiddenAvatars.map((presence) => (
-										<ListItem key={presence.clientID} sx={{ paddingY: 0.5 }}>
-											<ListItemAvatar>
-												<Avatar
-													sx={{
-														bgcolor: presence.presence.color,
-														width: 24,
-														height: 24,
-														fontSize: 12,
-													}}
-												>
-													{presence.presence.name[0]}
-												</Avatar>
-											</ListItemAvatar>
-											<ListItemText
-												primary={presence.presence.name}
-												primaryTypographyProps={{
-													variant: "body2",
-												}}
-											/>
-										</ListItem>
-									))}
-								</List>
-							</Paper>
-						</Popover>
+						<UserPresence presenceList={presenceList} />
 						{!editorState.shareRole && <ShareButton />}
 						<ThemeButton />
 					</Stack>
